debug meta

+++ classifier/features/__main__.py
@@ -57,7 +57,7 @@ class Main(Runner):
         return torch.stack([
             (
                 torch.tensor(
-                    sample[self.pipeline.col_names].astype(np.float64).values,
+                    sample[self.pipeline.col_names].values,
                     device=get_device()
                 )
                 .squeeze()
